backend/simple_nfts/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.7.2
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
base58==2.1.1
bitarray==1.2.2
certifi==2023.5.7
charset-normalizer==3.1.0
cytoolz==0.12.1
decorator==5.1.1
Django==4.2.2
django-cors-headers==4.1.0
django-debug-toolbar==4.1.0
django-extensions==3.2.3
django-metamask-auth==0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
eth-abi==2.2.0
eth-account==0.5.7
eth-hash==0.3.3
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
executing==1.2.0
frozenlist==1.3.3
hexbytes==0.3.1
idna==3.4
ipfshttpclient==0.8.0a2
ipython==8.14.0
jedi==0.18.2
jsonschema==3.2.0
lru-dict==1.2.0
matplotlib-inline==0.1.6
multiaddr==0.0.9
multidict==6.0.4
netaddr==0.8.0
parsimonious==0.8.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
prompt-toolkit==3.0.38
protobuf==3.20.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycryptodome==3.18.0
Pygments==2.15.1
PyJWT==2.7.0
pyrsistent==0.19.3
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
rlp==2.0.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
toolz==0.12.0
traitlets==5.9.0
typing_extensions==4.6.3
urllib3==2.0.3
varint==1.0.2
wcwidth==0.2.6
web3==5.28.0
websockets==9.1
yarl==1.9.2
